requirements.txt
requests
wxauto
PyYAML
